root: start deduplicating code
