+++ .bumpversion.cfg
@@ -27,8 +27,6 @@ values =
 
 [bumpversion:file:internal/constants/constants.go]
 
-[bumpversion:file:internal/outpost/version.go]
-
 [bumpversion:file:web/src/constants.ts]
 
 [bumpversion:file:website/docs/outposts/manual-deploy-docker-compose.md]
