+++ internal/constants/constants.go
@@ -1,3 +1,20 @@
 package constants
 
+import (
+	"fmt"
+	"os"
+)
+
 const VERSION = "2021.6.1-rc6"
+
+func BUILD() string {
+	build := os.Getenv("GIT_BUILD_HASH")
+	if build == "" {
+		return "tagged"
+	}
+	return build
+}
+
+func OutpostUserAgent() string {
+	return fmt.Sprintf("authentik-outpost@%s (%s)", VERSION, BUILD())
+}
